Ridisegnato la gestione delle CR;Eliminata obbligatorietà delle ore nelle attività;nella lista attività adesso consuntivazioni e richieste sono ordinate;sistemato campo ricerca, adesso ricorda il valore a video;aggiunto metodo per esecuzzione SQL con parametri;correzioni varie per migliorare funzionamento;

+++ routers/routeActivity.ts
@@ -6,7 +6,6 @@ import express, {Request, Response, Router} from "express";
 import CActivity from "../core/CActivity";
 import CWbs, {IWbs} from "../core/CWbs";
 import CCustomer from "../core/CCustomer";
-import CRequest from "../core/CRequest";
 import CWorkday from "../core/CWorkday";
 import {extraInfo, objectType} from "../core/CBase";
 
@@ -82,9 +81,9 @@ router.get("/:id", (request: Request, response: Response) => {
                wbs_id: w.id,
                customer_id: c.id,
                customer_description: c.description,
-               request: o.getRequest(),
                workday: o.getWorkday(),
                wbs_list: listOfWbs(o),
+               request_list: o.getRequest(),
                rdate: o.convertDate(new Date(), 3),
             }
          });
@@ -113,7 +112,7 @@ router.post("/:id", (request: Request, response: Response) => {
       o.wbs = parseInt(request.body.wbs);
       o.functional = request.body.functional;
       o.technical = request.body.technical;
-      o.hour = parseInt(request.body.hour);
+      o.hour = isNaN(parseInt(request.body.hour)) ? 0 : parseInt(request.body.hour);
       o.status = parseInt(request.body.status);
       o.note = request.body.note;
       o.save();
@@ -138,7 +137,6 @@ router.post("/:id", (request: Request, response: Response) => {
             wbs_id: w.id,
             customer_id: c.id,
             customer_description: c.description,
-            request: o.getRequest(),
             wbs_list: listOfWbs(o),
             error: e.message
          }
@@ -151,34 +149,49 @@ router.post("/:id", (request: Request, response: Response) => {
  */
 router.delete("/:id", (request: Request, response: Response) => {
    let o: CActivity;
-   let c: CRequest;
+   let w: CWbs;
+   let c: CCustomer;
 
    o = new CActivity();
-   c = new CRequest();
-   o.load(parseInt(request.params.id as string));
-   c.load(o.wbs);
-   if(!o.getRequest().length) {
-      o.delete();
+   w = new CWbs();
+   c = new CCustomer();
+
+   try {
+      o.load(parseInt(request.params.id as string));
+      w.load(o.wbs);
+      c.load(w.customer);
+      if(!o.getWorkday().length && !o.getRequest().length) {
+         o.delete();
+         response.redirect("/activity");
+      }
+      else
+         response.render("app", {
+            view: objectType.activity_details,
+            data: {
+               id: o.id,
+               internal_ref: o.internal_ref,
+               external_ref: o.external_ref,
+               type: o.type,
+               description: o.description,
+               functional: o.functional,
+               technical: o.technical,
+               hour: o.hour,
+               status: o.status,
+               note: o.note,
+               wbs_id: w.id,
+               customer_id: c.id,
+               customer_description: c.description,
+               workday: o.getWorkday(),
+               wbs_list: listOfWbs(o),
+               request_list: o.getRequest(),
+               rdate: o.convertDate(new Date(), 3),
+               error: "L'attività ha ancora oggetti assegnati."
+            }
+         });
+   }
+   catch(e) {
       response.redirect("/activity");
    }
-   else
-      response.render("app", {
-         view: objectType.activity_details,
-         data: {
-            id: o.id,
-            internal_ref: o.internal_ref,
-            external_ref: o.external_ref,
-            type: o.type,
-            description: o.description,
-            functional: o.functional,
-            technical: o.technical,
-            hour: o.hour,
-            status: o.status,
-            note: o.note,
-            wbs_list: listOfWbs(o),
-            error: "L'attività ha richieste di trasporto assegnate."
-         }
-      });
 });
 
 /**
@@ -206,24 +219,23 @@ router.post("/workday/:id", (request: Request, response: Response) => {
  * Inserimenti rapido di una richiesta di trasporto.
  */
 router.post("/request/:id", (request: Request, response: Response) => {
-   let o: CRequest;
-
-   o = new CRequest();
-
-   try {
-      o.activity = parseInt(request.params.id as string);
-      o.request = request.body.rrrequest;
-      o.type = parseInt(request.body.rrtype);
-      o.description = request.body.rrdescription;
-      o.owner = "";
-      o.date = new Date(request.body.rrdate);
-      o.note = "";
-      o.save();
-      o.transport();
-      o.saveTransport();
-      response.redirect(`/activity/${request.params.id}`);
-   }
-   catch(e) {
-      console.error(e);
-   }
+   // let o: CRequestOld;
+   //
+   // o = new CRequestOld();
+   //
+   // try {
+   //    o.activity = parseInt(request-old.params.id as string);
+   //    o.request-old = request-old.body.rrrequest;
+   //    o.type = parseInt(request-old.body.rrtype);
+   //    o.description = request-old.body.rrdescription;
+   //    o.owner = "";
+   //    o.date = new Date(request-old.body.rrdate);
+   //    o.note = "";
+   //    o.save();
+   //    o.transport();
+   //    response.redirect(`/activity/${request-old.params.id}`);
+   // }
+   // catch(e) {
+   //    console.error(e);
+   // }
 });
